refactor: split graphql-ide lib.rs into focused modules
Extract the monolithic lib.rs (8984 lines) into focused modules:

- database.rs: IdeDatabase, Salsa inputs, trait impls (110 lines)
- discovery.rs: File discovery types and functions (245 lines)
- host.rs: AnalysisHost struct and methods (728 lines)
- analysis.rs: Analysis snapshot and query methods (1187 lines)
- types.rs: Added SemanticToken*, FieldCoverageReport,
  ComplexityAnalysis and other POD types (1372 lines total)

lib.rs is now a thin re-export hub with module declarations and tests.

Addresses #628
