+++ crates/ide/src/lib.rs
@@ -38,19 +38,17 @@ mod analysis_host_isolation;
 #[cfg(test)]
 mod diagnostics_for_change_tests;
 
-use std::collections::HashMap;
-use std::sync::Arc;
-
-use parking_lot::RwLock;
-use salsa::Setter;
-
+// Infrastructure modules
+mod database;
+mod discovery;
 mod file_registry;
-pub use file_registry::FileRegistry;
-
-// New modular structure
 mod helpers;
 pub(crate) mod symbol;
 mod types;
+
+// Core modules
+mod analysis;
+mod host;
 
 // Feature modules
 mod code_lenses;
@@ -68,264 +66,31 @@ mod symbols;
 
 // Re-export types from the types module
 pub use types::{
-    CodeFix, CodeLens, CodeLensCommand, CodeLensInfo, CompletionItem, CompletionKind, Diagnostic,
-    DiagnosticSeverity, DocumentLoadResult, DocumentSymbol, FilePath, FoldingRange,
-    FoldingRangeKind, FragmentReference, FragmentUsage, HoverResult, InlayHint, InlayHintKind,
-    InsertTextFormat, Location, ParameterInformation, PendingIntrospection, Position,
-    ProjectStatus, Range, RenameResult, SchemaContentError, SchemaLoadResult, SchemaStats,
-    SelectionRange, SignatureHelp, SignatureInformation, SymbolKind, TextEdit, WorkspaceSymbol,
+    CodeFix, CodeLens, CodeLensCommand, CodeLensInfo, CompletionItem, CompletionKind,
+    ComplexityAnalysis, Diagnostic, DiagnosticSeverity, DocumentLoadResult, DocumentSymbol,
+    FieldComplexity, FieldCoverageReport, FieldUsageInfo, FilePath, FoldingRange, FoldingRangeKind,
+    FragmentReference, FragmentUsage, HoverResult, InlayHint, InlayHintKind, InsertTextFormat,
+    Location, ParameterInformation, PendingIntrospection, Position, ProjectStatus, Range,
+    RenameResult, SchemaContentError, SchemaLoadResult, SchemaStats, SelectionRange, SemanticToken,
+    SemanticTokenModifiers, SemanticTokenType, SignatureHelp, SignatureInformation, SymbolKind,
+    TextEdit, TypeCoverageInfo, WorkspaceSymbol,
 };
 
-// Re-export helpers for internal use
-use helpers::{adjust_range_for_line_offset, convert_diagnostic, offset_range_to_range};
+// Re-export file registry
+pub use file_registry::FileRegistry;
+
 // Re-export for use in symbol module and LSP
 pub use helpers::{path_to_file_uri, unwrap_type_to_name};
-
-use symbol::{find_fragment_definition_full_range, find_operation_definition_ranges};
 
 // Re-export database types that IDE layer needs
 pub use graphql_base_db::{DocumentKind, Language};
 
-/// Information about a loaded file from document discovery
-#[derive(Debug, Clone)]
-pub struct LoadedFile {
-    /// The file path (as a URI string)
-    pub path: FilePath,
-    /// The source language
-    pub language: Language,
-    /// The document kind
-    pub document_kind: DocumentKind,
-}
-
-/// File data that has been read from disk but not yet registered.
-/// Used to separate file I/O from lock acquisition.
-#[derive(Debug)]
-pub struct DiscoveredFile {
-    /// The file path (as a URI string)
-    pub path: FilePath,
-    /// The file content
-    pub content: String,
-    /// The source language
-    pub language: Language,
-    /// The document kind
-    pub document_kind: DocumentKind,
-}
-
-/// A content mismatch error found during file discovery.
-///
-/// This indicates a file's content doesn't match its expected `DocumentKind`
-/// based on which config pattern matched it.
-#[derive(Debug, Clone)]
-pub struct ContentMismatchError {
-    /// The pattern that matched this file
-    pub pattern: String,
-    /// Path to the file with mismatched content
-    pub file_path: std::path::PathBuf,
-    /// What kind was expected (based on config)
-    pub expected: graphql_config::FileType,
-    /// Names of definitions that don't belong
-    pub unexpected_definitions: Vec<String>,
-}
-
-/// Result of file discovery, containing both files and any validation errors.
-#[derive(Debug, Default)]
-pub struct FileDiscoveryResult {
-    /// Successfully discovered files
-    pub files: Vec<DiscoveredFile>,
-    /// Content mismatch errors found during discovery
-    pub errors: Vec<ContentMismatchError>,
-    /// Patterns that matched no files on disk
-    pub unmatched_patterns: Vec<String>,
-}
-
-impl FileDiscoveryResult {
-    /// Returns true if there are any content mismatch errors.
-    #[must_use]
-    pub fn has_errors(&self) -> bool {
-        !self.errors.is_empty()
-    }
-}
-
-/// Discover and read document files from config without requiring any locks.
-///
-/// This function performs all file I/O upfront so that lock acquisition
-/// for registration can be brief. Returns the file data ready for registration,
-/// along with any content mismatch errors found.
-///
-/// Files in the `documents` config are expected to contain executable definitions
-/// (operations, fragments). If schema definitions are found, an error is reported.
-pub fn discover_document_files(
-    config: &graphql_config::ProjectConfig,
-    workspace_path: &std::path::Path,
-) -> FileDiscoveryResult {
-    let Some(documents_config) = &config.documents else {
-        return FileDiscoveryResult::default();
-    };
-
-    let patterns: Vec<String> = documents_config
-        .patterns()
-        .into_iter()
-        .map(std::string::ToString::to_string)
-        .collect();
-
-    let mut result = FileDiscoveryResult::default();
-
-    for pattern in patterns {
-        // Skip negation patterns
-        if pattern.trim().starts_with('!') {
-            continue;
-        }
-
-        let expanded_patterns = expand_braces(&pattern);
-        let mut pattern_matched_any_files = false;
-
-        for expanded_pattern in expanded_patterns {
-            let full_pattern = workspace_path.join(&expanded_pattern);
-
-            match glob::glob(&full_pattern.display().to_string()) {
-                Ok(paths) => {
-                    for entry in paths {
-                        match entry {
-                            Ok(path) if path.is_file() => {
-                                // Skip node_modules
-                                if path.components().any(|c| c.as_os_str() == "node_modules") {
-                                    continue;
-                                }
-
-                                // Read file content
-                                match std::fs::read_to_string(&path) {
-                                    Ok(content) => {
-                                        let path_str = path.display().to_string();
-                                        let (language, document_kind) =
-                                            determine_document_file_kind(&path_str, &content);
-                                        let file_path = path_to_file_path(&path);
-
-                                        // Validate content matches expected kind (Executable)
-                                        // For TS/JS files, we need to extract GraphQL first
-                                        let graphql_content = if language.requires_extraction() {
-                                            // Extract and concatenate all GraphQL blocks
-                                            let config = graphql_extract::ExtractConfig::default();
-                                            graphql_extract::extract_from_source(
-                                                &content, language, &config, &path_str,
-                                            )
-                                            .unwrap_or_default()
-                                            .iter()
-                                            .map(|block| block.source.as_str())
-                                            .collect::<Vec<_>>()
-                                            .join("\n")
-                                        } else {
-                                            content.clone()
-                                        };
-
-                                        // Check for schema definitions in document files
-                                        if let Some(mismatch) =
-                                            graphql_syntax::validate_content_matches_kind(
-                                                &graphql_content,
-                                                DocumentKind::Executable,
-                                            )
-                                        {
-                                            let definitions = match mismatch {
-                                                graphql_syntax::ContentMismatch::ExpectedExecutableFoundSchema { definitions } => definitions,
-                                                graphql_syntax::ContentMismatch::ExpectedSchemaFoundExecutable { .. } => Vec::new(),
-                                            };
-                                            result.errors.push(ContentMismatchError {
-                                                pattern: pattern.clone(),
-                                                file_path: path.clone(),
-                                                expected: graphql_config::FileType::Document,
-                                                unexpected_definitions: definitions,
-                                            });
-                                        }
-
-                                        pattern_matched_any_files = true;
-                                        result.files.push(DiscoveredFile {
-                                            path: file_path,
-                                            content,
-                                            language,
-                                            document_kind,
-                                        });
-                                    }
-                                    Err(e) => {
-                                        tracing::warn!(
-                                            "Failed to read file {}: {}",
-                                            path.display(),
-                                            e
-                                        );
-                                    }
-                                }
-                            }
-                            Ok(_) => {}
-                            Err(e) => {
-                                tracing::warn!("Glob entry error: {}", e);
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    tracing::error!("Invalid glob pattern '{}': {}", expanded_pattern, e);
-                }
-            }
-        }
-
-        if !pattern_matched_any_files {
-            tracing::debug!("Document pattern matched no files: {}", pattern);
-            result.unmatched_patterns.push(pattern.clone());
-        }
-    }
-
-    result
-}
-
-/// Expand brace patterns like `{ts,tsx}` into multiple patterns
-///
-/// This is needed because the glob crate doesn't support brace expansion.
-/// For example, `**/*.{ts,tsx}` expands to `["**/*.ts", "**/*.tsx"]`.
-fn expand_braces(pattern: &str) -> Vec<String> {
-    if let Some(start) = pattern.find('{') {
-        if let Some(end) = pattern.find('}') {
-            let before = &pattern[..start];
-            let after = &pattern[end + 1..];
-            let options = &pattern[start + 1..end];
-
-            return options
-                .split(',')
-                .map(|opt| format!("{before}{opt}{after}"))
-                .collect();
-        }
-    }
-
-    vec![pattern.to_string()]
-}
-
-/// Check if a path has a given extension (case-insensitive)
-fn has_extension(path: &str, ext: &str) -> bool {
-    path.len() > ext.len()
-        && path.as_bytes()[path.len() - ext.len()..].eq_ignore_ascii_case(ext.as_bytes())
-}
-
-/// Determine `FileKind` for a document file based on its path.
-///
-/// This is used for files loaded from the `documents` configuration.
-/// - `.ts`/`.tsx` files → TypeScript
-/// - `.js`/`.jsx` files → JavaScript
-/// - `.graphql`/`.gql` files → `ExecutableGraphQL`
-///
-/// Note: Files from the `schema` configuration are always `Language::GraphQL, DocumentKind::Schema`,
-/// regardless of their extension.
-fn determine_document_file_kind(path: &str, _content: &str) -> (Language, DocumentKind) {
-    if has_extension(path, ".ts") || has_extension(path, ".tsx") {
-        (Language::TypeScript, DocumentKind::Executable)
-    } else if has_extension(path, ".js") || has_extension(path, ".jsx") {
-        (Language::JavaScript, DocumentKind::Executable)
-    } else {
-        (Language::GraphQL, DocumentKind::Executable)
-    }
-}
-
-/// Convert a filesystem path to a `FilePath` (URI format)
-fn path_to_file_path(path: &std::path::Path) -> FilePath {
-    let uri_string = path_to_file_uri(path);
-    FilePath::new(uri_string)
-}
+// Re-export core types
+pub use analysis::Analysis;
+pub use discovery::{
+    discover_document_files, ContentMismatchError, DiscoveredFile, FileDiscoveryResult, LoadedFile,
+};
+pub use host::AnalysisHost;
 
 #[cfg(test)]
 /// Helper for tests: extracts cursor position from a string with a `*` marker.
@@ -372,2316 +137,15 @@ fn extract_cursor(input: &str) -> (String, Position) {
     (result, Position::new(line, character))
 }
 
-// POD types are now defined in types.rs and re-exported above
-
-/// Semantic token type for syntax highlighting
-///
-/// These map to LSP semantic token types and provide rich syntax highlighting
-/// based on semantic analysis (e.g., knowing if a field is deprecated).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SemanticTokenType {
-    /// GraphQL type names (User, Post, etc.)
-    Type,
-    /// Field names in selection sets
-    Property,
-    /// Variables ($id, $limit)
-    Variable,
-    /// Fragment names
-    Function,
-    /// Enum values (ACTIVE, PENDING)
-    EnumMember,
-    /// Keywords (query, mutation, fragment, on)
-    Keyword,
-    /// String literals
-    String,
-    /// Number literals
-    Number,
-}
-
-impl SemanticTokenType {
-    /// Index into the legend (must match order in LSP capability registration)
-    #[must_use]
-    pub const fn index(self) -> u32 {
-        match self {
-            Self::Type => 0,
-            Self::Property => 1,
-            Self::Variable => 2,
-            Self::Function => 3,
-            Self::EnumMember => 4,
-            Self::Keyword => 5,
-            Self::String => 6,
-            Self::Number => 7,
-        }
-    }
-}
-
-/// Semantic token modifier for additional styling
-///
-/// These are combined as a bitmask and provide additional semantic information
-/// like deprecation status.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SemanticTokenModifiers(u32);
-
-impl SemanticTokenModifiers {
-    /// No modifiers
-    pub const NONE: Self = Self(0);
-    /// Element is deprecated (triggers strikethrough in editors)
-    pub const DEPRECATED: Self = Self(1 << 0);
-    /// Element is a definition site
-    pub const DEFINITION: Self = Self(1 << 1);
-
-    /// Create from raw bitmask
-    #[must_use]
-    pub const fn from_raw(raw: u32) -> Self {
-        Self(raw)
-    }
-
-    /// Get raw bitmask value
-    #[must_use]
-    pub const fn raw(self) -> u32 {
-        self.0
-    }
-
-    /// Combine modifiers
-    #[must_use]
-    pub const fn with(self, other: Self) -> Self {
-        Self(self.0 | other.0)
-    }
-
-    /// Check if a modifier is set
-    #[must_use]
-    pub const fn contains(self, other: Self) -> bool {
-        (self.0 & other.0) == other.0
-    }
-}
-
-/// A semantic token for syntax highlighting
-///
-/// Tokens are emitted in document order and converted to delta encoding
-/// by the LSP layer before being sent to the client.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SemanticToken {
-    /// Start position of the token
-    pub start: Position,
-    /// Length of the token in UTF-16 code units
-    pub length: u32,
-    /// Token type
-    pub token_type: SemanticTokenType,
-    /// Token modifiers (bitmask)
-    pub modifiers: SemanticTokenModifiers,
-}
-
-impl SemanticToken {
-    #[must_use]
-    pub const fn new(
-        start: Position,
-        length: u32,
-        token_type: SemanticTokenType,
-        modifiers: SemanticTokenModifiers,
-    ) -> Self {
-        Self {
-            start,
-            length,
-            token_type,
-            modifiers,
-        }
-    }
-}
-
-/// Field usage information for a single field
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FieldUsageInfo {
-    /// Number of operations that use this field
-    pub usage_count: usize,
-    /// Names of operations that use this field
-    pub operations: Vec<String>,
-}
-
-/// Coverage information for a single type
-#[derive(Debug, Clone, PartialEq)]
-pub struct TypeCoverageInfo {
-    /// Name of the type
-    pub type_name: String,
-    /// Total number of fields on this type
-    pub total_fields: usize,
-    /// Number of fields that are used in operations
-    pub used_fields: usize,
-}
-
-impl TypeCoverageInfo {
-    /// Calculate coverage as a percentage (0.0 to 100.0)
-    #[must_use]
-    #[allow(clippy::cast_precision_loss)]
-    pub fn coverage_percentage(&self) -> f64 {
-        if self.total_fields == 0 {
-            100.0
-        } else {
-            (self.used_fields as f64 / self.total_fields as f64) * 100.0
-        }
-    }
-}
-
-/// Field usage coverage report for an entire project
-#[derive(Debug, Clone, Default)]
-pub struct FieldCoverageReport {
-    /// Total number of fields in the schema
-    pub total_fields: usize,
-    /// Number of fields used in at least one operation
-    pub used_fields: usize,
-    /// Coverage by type
-    pub types: Vec<TypeCoverageInfo>,
-    /// Detailed field usages (`type_name`, `field_name`) -> usage info
-    pub field_usages: HashMap<(String, String), FieldUsageInfo>,
-}
-
-impl FieldCoverageReport {
-    /// Calculate overall coverage as a percentage (0.0 to 100.0)
-    #[must_use]
-    #[allow(clippy::cast_precision_loss)]
-    pub fn coverage_percentage(&self) -> f64 {
-        if self.total_fields == 0 {
-            100.0
-        } else {
-            (self.used_fields as f64 / self.total_fields as f64) * 100.0
-        }
-    }
-
-    /// Get all unused fields as (`type_name`, `field_name`) tuples
-    #[must_use]
-    pub fn unused_fields(&self) -> Vec<(String, String)> {
-        self.field_usages
-            .iter()
-            .filter(|(_, info)| info.usage_count == 0)
-            .map(|((type_name, field_name), _)| (type_name.clone(), field_name.clone()))
-            .collect()
-    }
-}
-
-impl From<Arc<graphql_analysis::FieldCoverageReport>> for FieldCoverageReport {
-    fn from(report: Arc<graphql_analysis::FieldCoverageReport>) -> Self {
-        let types: Vec<TypeCoverageInfo> = report
-            .type_coverage
-            .iter()
-            .map(|(name, coverage)| TypeCoverageInfo {
-                type_name: name.to_string(),
-                total_fields: coverage.total_fields,
-                used_fields: coverage.used_fields,
-            })
-            .collect();
-
-        let field_usages: HashMap<(String, String), FieldUsageInfo> = report
-            .field_usages
-            .iter()
-            .map(|((type_name, field_name), usage)| {
-                (
-                    (type_name.to_string(), field_name.to_string()),
-                    FieldUsageInfo {
-                        usage_count: usage.usage_count,
-                        operations: usage.operations.iter().map(ToString::to_string).collect(),
-                    },
-                )
-            })
-            .collect();
-
-        Self {
-            total_fields: report.total_fields,
-            used_fields: report.used_fields,
-            types,
-            field_usages,
-        }
-    }
-}
-
-/// Per-field complexity breakdown
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FieldComplexity {
-    /// Field path from root (e.g., "posts.author.name")
-    pub path: String,
-    /// Field name
-    pub name: String,
-    /// Complexity score for this field
-    pub complexity: u32,
-    /// List multiplier (for list fields like `[Post!]!`)
-    pub multiplier: u32,
-    /// Depth level (0 = root level)
-    pub depth: u32,
-    /// Whether this is a connection pattern (edges/nodes pagination)
-    pub is_connection: bool,
-    /// Warning message if any (e.g., nested pagination)
-    pub warning: Option<String>,
-}
-
-impl FieldComplexity {
-    pub fn new(path: impl Into<String>, name: impl Into<String>, complexity: u32) -> Self {
-        Self {
-            path: path.into(),
-            name: name.into(),
-            complexity,
-            multiplier: 1,
-            depth: 0,
-            is_connection: false,
-            warning: None,
-        }
-    }
-
-    #[must_use]
-    pub const fn with_multiplier(mut self, multiplier: u32) -> Self {
-        self.multiplier = multiplier;
-        self
-    }
-
-    #[must_use]
-    pub const fn with_depth(mut self, depth: u32) -> Self {
-        self.depth = depth;
-        self
-    }
-
-    #[must_use]
-    pub const fn with_connection(mut self, is_connection: bool) -> Self {
-        self.is_connection = is_connection;
-        self
-    }
-
-    #[must_use]
-    pub fn with_warning(mut self, warning: impl Into<String>) -> Self {
-        self.warning = Some(warning.into());
-        self
-    }
-}
-
-/// Complexity analysis result for an operation
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ComplexityAnalysis {
-    /// Operation name (or "<anonymous>" for unnamed operations)
-    pub operation_name: String,
-    /// Operation type (query, mutation, subscription)
-    pub operation_type: String,
-    /// Total calculated complexity score
-    pub total_complexity: u32,
-    /// Maximum selection depth
-    pub depth: u32,
-    /// Per-field complexity breakdown
-    pub breakdown: Vec<FieldComplexity>,
-    /// Warnings about potential issues (nested pagination, etc.)
-    pub warnings: Vec<String>,
-    /// File path containing this operation
-    pub file: FilePath,
-    /// Range of the operation in the file
-    pub range: Range,
-}
-
-impl ComplexityAnalysis {
-    pub fn new(
-        operation_name: impl Into<String>,
-        operation_type: impl Into<String>,
-        file: FilePath,
-        range: Range,
-    ) -> Self {
-        Self {
-            operation_name: operation_name.into(),
-            operation_type: operation_type.into(),
-            total_complexity: 0,
-            depth: 0,
-            breakdown: Vec::new(),
-            warnings: Vec::new(),
-            file,
-            range,
-        }
-    }
-}
-
-/// Input: Lint configuration
-///
-/// This is a Salsa input so that config changes properly invalidate dependent queries.
-/// Wrapping in Arc allows queries to access the config without cloning the entire config object.
-///
-/// Using Salsa inputs instead of `Arc<RwLock<...>>` ensures:
-/// - Proper dependency tracking (Salsa knows which queries depend on config)
-/// - Automatic invalidation (only config-dependent queries re-run on changes)
-/// - No deadlock risk (Salsa manages all locking internally)
-/// - Snapshot isolation (config is immutable in Analysis snapshots)
-#[salsa::input]
-struct LintConfigInput {
-    pub config: Arc<graphql_linter::LintConfig>,
-}
-
-/// Input: Extract configuration for TypeScript/JavaScript extraction
-///
-/// This is a Salsa input so that config changes properly invalidate dependent queries.
-///
-/// Using Salsa inputs instead of `Arc<RwLock<...>>` ensures:
-/// - Proper dependency tracking (Salsa knows which queries depend on config)
-/// - Automatic invalidation (only config-dependent queries re-run on changes)
-/// - No deadlock risk (Salsa manages all locking internally)
-/// - Snapshot isolation (config is immutable in Analysis snapshots)
-#[salsa::input]
-struct ExtractConfigInput {
-    pub config: Arc<graphql_extract::ExtractConfig>,
-}
-
-/// Custom database that implements config traits
-///
-/// All configuration is now stored as Salsa inputs (`LintConfigInput`, `ExtractConfigInput`,
-/// and `ProjectFiles`) instead of `Arc<RwLock<...>>` wrappers. This allows Salsa to properly
-/// track config dependencies and only invalidate affected queries when inputs change.
-///
-/// Queries can access `project_files` via `db.project_files()` and Salsa will automatically
-/// track dependencies when the query calls getters like `project_files.schema_file_ids(db)`.
-#[salsa::db]
-#[derive(Clone)]
-struct IdeDatabase {
-    storage: salsa::Storage<Self>,
-    lint_config_input: Option<LintConfigInput>,
-    extract_config_input: Option<ExtractConfigInput>,
-    /// Project files input - stores the current `ProjectFiles` Salsa input directly.
-    /// Unlike the old `Arc<RwLock<...>>` approach, this enables proper Salsa dependency
-    /// tracking: queries that call `db.project_files()` and then access fields like
-    /// `project_files.schema_file_ids(db)` will have their dependencies tracked.
-    project_files_input: Option<graphql_base_db::ProjectFiles>,
-}
-
-impl Default for IdeDatabase {
-    fn default() -> Self {
-        let mut db = Self {
-            storage: salsa::Storage::new(Some(Box::new(|event: salsa::Event| match event.kind {
-                salsa::EventKind::WillExecute { database_key, .. } => {
-                    tracing::debug!("query cache miss (executing): {database_key:?}");
-                }
-                salsa::EventKind::DidValidateMemoizedValue { database_key } => {
-                    tracing::debug!("query cache hit (memoized): {database_key:?}");
-                }
-                _ => {}
-            }))),
-            lint_config_input: None,
-            extract_config_input: None,
-            project_files_input: None,
-        };
-
-        // Initialize with default configs as Salsa inputs
-        db.lint_config_input = Some(LintConfigInput::new(
-            &db,
-            Arc::new(graphql_linter::LintConfig::default()),
-        ));
-        db.extract_config_input = Some(ExtractConfigInput::new(
-            &db,
-            Arc::new(graphql_extract::ExtractConfig::default()),
-        ));
-
-        db
-    }
-}
-
-#[salsa::db]
-impl salsa::Database for IdeDatabase {}
-
-#[salsa::db]
-impl graphql_syntax::GraphQLSyntaxDatabase for IdeDatabase {
-    fn extract_config(&self) -> Option<Arc<graphql_extract::ExtractConfig>> {
-        self.extract_config_input
-            .map(|input| input.config(self).clone())
-    }
-}
-
-#[salsa::db]
-impl graphql_hir::GraphQLHirDatabase for IdeDatabase {
-    fn project_files(&self) -> Option<graphql_base_db::ProjectFiles> {
-        self.project_files_input
-    }
-}
-
-#[salsa::db]
-impl graphql_analysis::GraphQLAnalysisDatabase for IdeDatabase {
-    fn lint_config(&self) -> Arc<graphql_linter::LintConfig> {
-        self.lint_config_input.map_or_else(
-            || Arc::new(graphql_linter::LintConfig::default()),
-            |input| input.config(self).clone(),
-        )
-    }
-}
-
-/// The main analysis host
-///
-/// This is the entry point for all IDE features. It owns the database and
-/// provides methods to apply changes and create snapshots for analysis.
-///
-/// # Snapshot Lifecycle (Important!)
-///
-/// This follows Salsa's single-writer, multi-reader model:
-/// - Call [`snapshot()`](Self::snapshot) to create an immutable [`Analysis`] for queries
-/// - **All snapshots must be dropped before calling any mutating method**
-/// - Failing to drop snapshots before mutation will cause a hang/deadlock
-///
-/// ```ignore
-/// // CORRECT: Scope snapshots so they're dropped before mutation
-/// let result = {
-///     let snapshot = host.snapshot();
-///     snapshot.diagnostics(&file)
-/// }; // snapshot dropped here
-/// host.add_file(&file, new_content, kind); // Safe: no snapshots exist
-///
-/// // WRONG: Holding snapshot across mutation
-/// let snapshot = host.snapshot();
-/// let result = snapshot.diagnostics(&file);
-/// host.add_file(&file, new_content, kind); // HANGS: snapshot still alive!
-/// ```
-pub struct AnalysisHost {
-    db: IdeDatabase,
-    /// File registry for mapping paths to file IDs
-    /// Wrapped in Arc<RwLock> so snapshots can share it
-    registry: Arc<RwLock<FileRegistry>>,
-}
-
-impl AnalysisHost {
-    /// Create a new analysis host with a default database
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            db: IdeDatabase::default(),
-            registry: Arc::new(RwLock::new(FileRegistry::new())),
-        }
-    }
-
-    /// Add or update a file in the host
-    ///
-    /// This is a convenience method for adding files to the registry and database.
-    ///
-    /// Returns `true` if this is a new file, `false` if it's an update to an existing file.
-    ///
-    /// **IMPORTANT**: Only call `rebuild_project_files()` when this returns `true` (new file).
-    /// Content-only updates do NOT require rebuilding the project index.
-    pub fn add_file(
-        &mut self,
-        path: &FilePath,
-        content: &str,
-        language: Language,
-        document_kind: DocumentKind,
-    ) -> bool {
-        let mut registry = self.registry.write();
-        let (_, _, _, is_new) =
-            registry.add_file(&mut self.db, path, content, language, document_kind);
-        is_new
-    }
-
-    /// Batch-add pre-discovered files to the host.
-    ///
-    /// This is more efficient than calling `add_file` in a loop because
-    /// file I/O has already been done. The lock is only held briefly
-    /// for registration, not during disk reads.
-    ///
-    /// Returns the list of `LoadedFile` structs for building indexes.
-    pub fn add_discovered_files(&mut self, files: &[DiscoveredFile]) -> Vec<LoadedFile> {
-        let mut registry = self.registry.write();
-        let mut loaded = Vec::with_capacity(files.len());
-
-        for file in files {
-            registry.add_file(
-                &mut self.db,
-                &file.path,
-                &file.content,
-                file.language,
-                file.document_kind,
-            );
-            loaded.push(LoadedFile {
-                path: file.path.clone(),
-                language: file.language,
-                document_kind: file.document_kind,
-            });
-        }
-
-        loaded
-    }
-
-    /// Rebuild the `ProjectFiles` index after adding/removing files
-    ///
-    /// This should be called after batch adding files to avoid O(n²) performance.
-    /// It's relatively expensive as it iterates through all files, so avoid calling
-    /// it in a loop.
-    ///
-    /// This method also syncs the `ProjectFiles` to the database so queries can
-    /// access it via `db.project_files()`.
-    pub fn rebuild_project_files(&mut self) {
-        let mut registry = self.registry.write();
-        registry.rebuild_project_files(&mut self.db);
-
-        // Sync project_files from registry to database
-        // This enables queries to access project_files via db.project_files()
-        self.db.project_files_input = registry.project_files();
-    }
-
-    /// Add multiple files in batch, then rebuild the project index once
-    ///
-    /// This is the recommended way to load multiple files at once. It:
-    /// 1. Adds all files to the registry without rebuilding
-    /// 2. Rebuilds the project index once at the end
-    ///
-    /// This is O(n) instead of O(n²) compared to calling `add_file` + `rebuild_project_files`
-    /// for each file individually.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use graphql_ide::{AnalysisHost, FilePath, Language, DocumentKind};
-    ///
-    /// let mut host = AnalysisHost::new();
-    /// let files = vec![
-    ///     (FilePath::new("file:///schema.graphql"), "type Query { hello: String }", Language::GraphQL, DocumentKind::Schema),
-    ///     (FilePath::new("file:///query.graphql"), "query { hello }", Language::GraphQL, DocumentKind::Executable),
-    /// ];
-    /// host.add_files_batch(&files);
-    /// ```
-    pub fn add_files_batch(&mut self, files: &[(FilePath, &str, Language, DocumentKind)]) {
-        let mut registry = self.registry.write();
-        let mut any_new = false;
-
-        for (path, content, language, document_kind) in files {
-            let (_, _, _, is_new) =
-                registry.add_file(&mut self.db, path, content, *language, *document_kind);
-            any_new = any_new || is_new;
-        }
-
-        // Only rebuild if at least one file was new
-        if any_new {
-            registry.rebuild_project_files(&mut self.db);
-            // Sync project_files from registry to database
-            self.db.project_files_input = registry.project_files();
-        }
-    }
-
-    /// Update a file and optionally create a snapshot in a single lock acquisition
-    ///
-    /// This is optimized for the common case of editing an existing file. It:
-    /// 1. Updates the file content (cheap operation)
-    /// 2. Returns a snapshot for immediate analysis
-    ///
-    /// This avoids the overhead of multiple lock acquisitions per keystroke.
-    /// For new files, call `add_file()` followed by `rebuild_project_files()` instead.
-    ///
-    /// Returns `(is_new_file, Analysis)` tuple.
-    pub fn update_file_and_snapshot(
-        &mut self,
-        path: &FilePath,
-        content: &str,
-        language: Language,
-        document_kind: DocumentKind,
-    ) -> (bool, Analysis) {
-        // Single lock acquisition for both operations
-        let mut registry = self.registry.write();
-        let (_, _, _, is_new) =
-            registry.add_file(&mut self.db, path, content, language, document_kind);
-
-        // If this is a new file, rebuild the index before creating snapshot
-        // This also syncs project_files to self.db.project_files_input
-        if is_new {
-            registry.rebuild_project_files(&mut self.db);
-            // Sync project_files from registry to database
-            self.db.project_files_input = registry.project_files();
-        }
-
-        let project_files = self.db.project_files_input;
-        // Release the lock before creating the snapshot (no longer needed)
-        drop(registry);
-
-        let snapshot = Analysis {
-            db: self.db.clone(),
-            registry: Arc::clone(&self.registry),
-            project_files,
-        };
-
-        (is_new, snapshot)
-    }
-
-    /// Check if a file exists in this host's registry
-    #[must_use]
-    pub fn contains_file(&self, path: &FilePath) -> bool {
-        let registry = self.registry.read();
-        registry.get_file_id(path).is_some()
-    }
-
-    /// Remove a file from the host
-    pub fn remove_file(&mut self, path: &FilePath) {
-        let mut registry = self.registry.write();
-        if let Some(file_id) = registry.get_file_id(path) {
-            registry.remove_file(file_id);
-            registry.rebuild_project_files(&mut self.db);
-        }
-    }
-
-    /// Load schema files from a project configuration
-    ///
-    /// This method:
-    /// - Always includes Apollo Client built-in directives
-    /// - Loads schema files from local paths (single file, multiple files, glob patterns)
-    /// - Supports TypeScript/JavaScript files with embedded GraphQL schemas
-    /// - Collects remote introspection configs for async fetching by the caller
-    ///
-    /// Returns a [`SchemaLoadResult`] containing the count of loaded files and any
-    /// pending introspection configurations that require async fetching.
-    pub fn load_schemas_from_config(
-        &mut self,
-        config: &graphql_config::ProjectConfig,
-        base_dir: &std::path::Path,
-    ) -> anyhow::Result<SchemaLoadResult> {
-        const SCHEMA_BUILTINS: &str = include_str!("schema_builtins.graphql");
-        const APOLLO_CLIENT_BUILTINS: &str = include_str!("apollo_client_builtins.graphql");
-        const RELAY_CLIENT_BUILTINS: &str = include_str!("relay_client_builtins.graphql");
-
-        // Always include GraphQL spec built-in directives first (e.g., @oneOf)
-        self.add_file(
-            &FilePath::new("schema_builtins.graphql".to_string()),
-            SCHEMA_BUILTINS,
-            Language::GraphQL,
-            DocumentKind::Schema,
-        );
-        let mut count = 1;
-
-        // Include client-specific built-in directives based on config
-        match config.client() {
-            Some(graphql_config::ClientConfig::Apollo) => {
-                self.add_file(
-                    &FilePath::new("client_builtins.graphql".to_string()),
-                    APOLLO_CLIENT_BUILTINS,
-                    Language::GraphQL,
-                    DocumentKind::Schema,
-                );
-                count += 1;
-            }
-            Some(graphql_config::ClientConfig::Relay) => {
-                self.add_file(
-                    &FilePath::new("client_builtins.graphql".to_string()),
-                    RELAY_CLIENT_BUILTINS,
-                    Language::GraphQL,
-                    DocumentKind::Schema,
-                );
-                count += 1;
-            }
-            Some(graphql_config::ClientConfig::None) | None => {
-                // No client directives
-            }
-        }
-        let mut loaded_paths = Vec::new();
-        let mut pending_introspections = Vec::new();
-        let mut content_errors = Vec::new();
-        let mut unmatched_patterns = Vec::new();
-
-        let patterns: Vec<String> = match &config.schema {
-            graphql_config::SchemaConfig::Path(s) => vec![s.clone()],
-            graphql_config::SchemaConfig::Paths(arr) => arr.clone(),
-            graphql_config::SchemaConfig::Introspection(introspection) => {
-                // Collect introspection config for async fetching by the caller
-                tracing::info!(
-                    "Found remote schema introspection config: {}",
-                    introspection.url
-                );
-                pending_introspections.push(PendingIntrospection::from_config(introspection));
-                vec![]
-            }
-        };
-
-        for pattern in patterns {
-            // Collect URL patterns as pending introspections for async fetching
-            if pattern.starts_with("http://") || pattern.starts_with("https://") {
-                tracing::info!("Found remote schema URL: {}", pattern);
-                pending_introspections.push(PendingIntrospection {
-                    url: pattern,
-                    headers: None,
-                    timeout: None,
-                    retry: None,
-                });
-                continue;
-            }
-
-            // Treat as file glob pattern
-            let full_pattern = base_dir.join(&pattern).display().to_string();
-
-            match glob::glob(&full_pattern) {
-                Ok(paths) => {
-                    let mut pattern_matched_files = false;
-                    for entry in paths.flatten() {
-                        if entry.is_file() {
-                            pattern_matched_files = true;
-                            match std::fs::read_to_string(&entry) {
-                                Ok(content) => {
-                                    let file_uri = path_to_file_uri(&entry);
-                                    let language = graphql_extract::Language::from_path(&entry);
-
-                                    // Check if this is a TS/JS file that needs extraction
-                                    if let Some(lang) = language {
-                                        if lang.requires_extraction() {
-                                            // Extract GraphQL from TS/JS file
-                                            let extract_config = self.get_extract_config();
-                                            match graphql_extract::extract_from_source(
-                                                &content,
-                                                lang,
-                                                &extract_config,
-                                                &file_uri,
-                                            ) {
-                                                Ok(blocks) => {
-                                                    // Validate all blocks for executable definitions
-                                                    let all_sources: String = blocks
-                                                        .iter()
-                                                        .map(|b| b.source.as_str())
-                                                        .collect::<Vec<_>>()
-                                                        .join("\n");
-                                                    if let Some(mismatch) = graphql_syntax::validate_content_matches_kind(
-                                                        &all_sources,
-                                                        DocumentKind::Schema,
-                                                    ) {
-                                                        let definitions = match mismatch {
-                                                            graphql_syntax::ContentMismatch::ExpectedSchemaFoundExecutable { definitions } => definitions,
-                                                            graphql_syntax::ContentMismatch::ExpectedExecutableFoundSchema { .. } => Vec::new(),
-                                                        };
-                                                        content_errors.push(SchemaContentError {
-                                                            pattern: pattern.clone(),
-                                                            file_path: entry.clone(),
-                                                            unexpected_definitions: definitions,
-                                                        });
-                                                    }
-
-                                                    if blocks.len() == 1 {
-                                                        // Single block: store original TS/JS content
-                                                        // so the syntax crate can handle extraction
-                                                        // with proper line offsets
-                                                        self.add_file(
-                                                            &FilePath::new(file_uri.clone()),
-                                                            &content,
-                                                            lang,
-                                                            DocumentKind::Schema,
-                                                        );
-                                                        count += 1;
-                                                    } else {
-                                                        // Multiple blocks: create separate entries
-                                                        // with line range URIs for each block
-                                                        for block in &blocks {
-                                                            let start_line =
-                                                                block.location.range.start.line + 1;
-                                                            let end_line =
-                                                                block.location.range.end.line + 1;
-                                                            let block_uri = format!(
-                                                                "{file_uri}#L{start_line}-L{end_line}"
-                                                            );
-
-                                                            self.add_file(
-                                                                &FilePath::new(block_uri),
-                                                                &block.source,
-                                                                Language::GraphQL,
-                                                                DocumentKind::Schema,
-                                                            );
-                                                            count += 1;
-                                                        }
-                                                    }
-                                                    if blocks.is_empty() {
-                                                        tracing::debug!(
-                                                            "No GraphQL blocks found in {}",
-                                                            entry.display()
-                                                        );
-                                                    } else {
-                                                        loaded_paths.push(entry.clone());
-                                                    }
-                                                }
-                                                Err(e) => {
-                                                    tracing::warn!(
-                                                        "Failed to extract GraphQL from {}: {}",
-                                                        entry.display(),
-                                                        e
-                                                    );
-                                                }
-                                            }
-                                            continue;
-                                        }
-                                    }
-
-                                    // Pure GraphQL file - validate and add
-                                    // Check for executable definitions (operations/fragments)
-                                    if let Some(mismatch) =
-                                        graphql_syntax::validate_content_matches_kind(
-                                            &content,
-                                            DocumentKind::Schema,
-                                        )
-                                    {
-                                        let definitions = match mismatch {
-                                            graphql_syntax::ContentMismatch::ExpectedSchemaFoundExecutable { definitions } => definitions,
-                                            graphql_syntax::ContentMismatch::ExpectedExecutableFoundSchema { .. } => Vec::new(),
-                                        };
-                                        content_errors.push(SchemaContentError {
-                                            pattern: pattern.clone(),
-                                            file_path: entry.clone(),
-                                            unexpected_definitions: definitions,
-                                        });
-                                    }
-
-                                    self.add_file(
-                                        &FilePath::new(file_uri),
-                                        &content,
-                                        Language::GraphQL,
-                                        DocumentKind::Schema,
-                                    );
-                                    loaded_paths.push(entry.clone());
-                                    count += 1;
-                                }
-                                Err(e) => {
-                                    let path_display = entry.display().to_string();
-                                    tracing::error!(
-                                        "Failed to read schema file {path_display}: {e}"
-                                    );
-                                    return Err(anyhow::anyhow!(
-                                        "Failed to read schema file {path_display}: {e}"
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                    if !pattern_matched_files {
-                        tracing::debug!(
-                            "Schema pattern matched no files: {} (expanded: {})",
-                            pattern,
-                            full_pattern
-                        );
-                        unmatched_patterns.push(pattern.clone());
-                    }
-                }
-                Err(e) => {
-                    tracing::error!("Failed to expand glob pattern {full_pattern}: {e}");
-                    return Err(anyhow::anyhow!(
-                        "Failed to expand glob pattern {full_pattern}: {e}"
-                    ));
-                }
-            }
-        }
-
-        tracing::info!(
-            "Loaded {} schema file(s) ({} paths tracked), {} pending introspection(s)",
-            count,
-            loaded_paths.len(),
-            pending_introspections.len()
-        );
-        Ok(SchemaLoadResult {
-            loaded_count: count,
-            loaded_paths,
-            pending_introspections,
-            content_errors,
-            unmatched_patterns,
-        })
-    }
-
-    /// Add an introspected schema as a virtual file.
-    ///
-    /// This method registers a schema fetched via introspection using a virtual URI
-    /// in the format `schema://<host>/<path>/schema.graphql`. The `.graphql`
-    /// extension ensures editors recognize the file for syntax highlighting and
-    /// language features.
-    ///
-    /// # Arguments
-    ///
-    /// * `url` - The original GraphQL endpoint URL (used to generate the virtual URI)
-    /// * `sdl` - The schema SDL obtained from introspection
-    ///
-    /// # Returns
-    ///
-    /// The virtual file URI used to register the schema.
-    pub fn add_introspected_schema(&mut self, url: &str, sdl: &str) -> String {
-        let virtual_uri = format!(
-            "schema://{}/schema.graphql",
-            url.trim_start_matches("https://")
-                .trim_start_matches("http://")
-        );
-        tracing::info!("Adding introspected schema from {} as {}", url, virtual_uri);
-        self.add_file(
-            &FilePath::new(virtual_uri.clone()),
-            sdl,
-            Language::GraphQL,
-            DocumentKind::Schema,
-        );
-        virtual_uri
-    }
-
-    /// Set the lint configuration for the project
-    ///
-    /// This properly invalidates all queries that depend on lint config via Salsa's
-    /// dependency tracking. Only lint-dependent queries will re-run when config changes.
-    pub fn set_lint_config(&mut self, config: graphql_linter::LintConfig) {
-        if let Some(input) = self.db.lint_config_input {
-            input.set_config(&mut self.db).to(Arc::new(config));
-        } else {
-            let input = LintConfigInput::new(&self.db, Arc::new(config));
-            self.db.lint_config_input = Some(input);
-        }
-    }
-
-    /// Set the extract configuration for the project
-    ///
-    /// This properly invalidates all queries that depend on extract config via Salsa's
-    /// dependency tracking. Only extract-dependent queries will re-run when config changes.
-    pub fn set_extract_config(&mut self, config: graphql_extract::ExtractConfig) {
-        if let Some(input) = self.db.extract_config_input {
-            input.set_config(&mut self.db).to(Arc::new(config));
-        } else {
-            let input = ExtractConfigInput::new(&self.db, Arc::new(config));
-            self.db.extract_config_input = Some(input);
-        }
-    }
-
-    /// Get the extract configuration for the project
-    pub fn get_extract_config(&self) -> graphql_extract::ExtractConfig {
-        self.db
-            .extract_config_input
-            .map(|input| (*input.config(&self.db)).clone())
-            .unwrap_or_default()
-    }
-
-    /// Load document files from a project configuration
-    ///
-    /// This method handles:
-    /// - Glob pattern expansion (including brace expansion like `{ts,tsx}`)
-    /// - File reading and content extraction
-    /// - File kind determination based on extension
-    /// - Batch file registration (more efficient than individual `add_file` calls)
-    ///
-    /// Returns information about loaded files for indexing purposes.
-    ///
-    /// # Arguments
-    ///
-    /// * `config` - The project configuration containing document patterns
-    /// * `workspace_path` - The base directory for glob pattern resolution
-    ///
-    /// # Returns
-    ///
-    /// A vector of `LoadedFile` structs containing file paths and metadata.
-    /// The caller can use this information to build file-to-project indexes.
-    pub fn load_documents_from_config(
-        &mut self,
-        config: &graphql_config::ProjectConfig,
-        workspace_path: &std::path::Path,
-    ) -> (Vec<LoadedFile>, DocumentLoadResult) {
-        let Some(documents_config) = &config.documents else {
-            return (Vec::new(), DocumentLoadResult::default());
-        };
-
-        let patterns: Vec<String> = documents_config
-            .patterns()
-            .into_iter()
-            .map(std::string::ToString::to_string)
-            .collect();
-
-        let mut loaded_files: Vec<LoadedFile> = Vec::new();
-        let mut files_to_add: Vec<(FilePath, String, Language, DocumentKind)> = Vec::new();
-        let mut unmatched_patterns: Vec<String> = Vec::new();
-
-        for pattern in patterns {
-            // Skip negation patterns
-            if pattern.trim().starts_with('!') {
-                continue;
-            }
-
-            let expanded_patterns = expand_braces(&pattern);
-            let mut pattern_matched_any_files = false;
-
-            for expanded_pattern in expanded_patterns {
-                let full_pattern = workspace_path.join(&expanded_pattern);
-
-                match glob::glob(&full_pattern.display().to_string()) {
-                    Ok(paths) => {
-                        for entry in paths {
-                            match entry {
-                                Ok(path) if path.is_file() => {
-                                    // Skip node_modules
-                                    if path.components().any(|c| c.as_os_str() == "node_modules") {
-                                        continue;
-                                    }
-
-                                    // Read file content
-                                    match std::fs::read_to_string(&path) {
-                                        Ok(content) => {
-                                            let path_str = path.display().to_string();
-                                            let (language, document_kind) =
-                                                determine_document_file_kind(&path_str, &content);
-
-                                            let file_path = path_to_file_path(&path);
-
-                                            pattern_matched_any_files = true;
-                                            loaded_files.push(LoadedFile {
-                                                path: file_path.clone(),
-                                                language,
-                                                document_kind,
-                                            });
-
-                                            files_to_add.push((
-                                                file_path,
-                                                content,
-                                                language,
-                                                document_kind,
-                                            ));
-                                        }
-                                        Err(e) => {
-                                            tracing::warn!(
-                                                "Failed to read file {}: {}",
-                                                path.display(),
-                                                e
-                                            );
-                                        }
-                                    }
-                                }
-                                Ok(_) => {}
-                                Err(e) => {
-                                    tracing::warn!("Glob entry error: {}", e);
-                                }
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        tracing::error!("Invalid glob pattern '{}': {}", expanded_pattern, e);
-                    }
-                }
-            }
-
-            if !pattern_matched_any_files {
-                tracing::debug!("Document pattern matched no files: {}", pattern);
-                unmatched_patterns.push(pattern.clone());
-            }
-        }
-
-        // Batch add all files using add_files_batch for O(n) performance
-        // Convert owned strings to borrowed for the batch API
-        let batch_refs: Vec<(FilePath, &str, Language, DocumentKind)> = files_to_add
-            .iter()
-            .map(|(path, content, language, document_kind)| {
-                (path.clone(), content.as_str(), *language, *document_kind)
-            })
-            .collect();
-        self.add_files_batch(&batch_refs);
-
-        let result = DocumentLoadResult {
-            loaded_count: loaded_files.len(),
-            unmatched_patterns,
-        };
-
-        (loaded_files, result)
-    }
-
-    /// Iterate over all files in the host
-    ///
-    /// Returns an iterator of `FilePath` for all registered files.
-    pub fn files(&self) -> Vec<FilePath> {
-        let registry = self.registry.read();
-        registry
-            .all_file_ids()
-            .into_iter()
-            .filter_map(|file_id| registry.get_path(file_id))
-            .collect()
-    }
-
-    /// Get an immutable snapshot for analysis
-    ///
-    /// This snapshot can be used from multiple threads and provides all IDE features.
-    /// It's cheap to create and clone (`RootDatabase` implements Clone via salsa).
-    ///
-    /// # Lifecycle Warning
-    ///
-    /// The returned `Analysis` **must be dropped before calling any mutating method**
-    /// on this `AnalysisHost`. This is required by Salsa's single-writer model.
-    /// See the struct-level documentation for details and examples.
-    pub fn snapshot(&self) -> Analysis {
-        // project_files is already synced to the database in rebuild_project_files()
-        // Queries access it via db.project_files()
-        let project_files = self.db.project_files_input;
-
-        if let Some(ref pf) = project_files {
-            let doc_count = pf.document_file_ids(&self.db).ids(&self.db).len();
-            let schema_count = pf.schema_file_ids(&self.db).ids(&self.db).len();
-            tracing::debug!(
-                "Snapshot project_files: {} schema files, {} document files",
-                schema_count,
-                doc_count
-            );
-        } else {
-            tracing::warn!("Snapshot project_files is None!");
-        }
-
-        Analysis {
-            db: self.db.clone(),
-            registry: Arc::clone(&self.registry),
-            project_files,
-        }
-    }
-}
-
-impl Default for AnalysisHost {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Immutable snapshot of the analysis state
-///
-/// Can be cheaply cloned and used from multiple threads.
-/// All IDE feature queries go through this.
-///
-/// # Lifecycle Warning
-///
-/// This snapshot shares Salsa storage with its parent [`AnalysisHost`].
-/// **You must drop all `Analysis` instances before calling any mutating method**
-/// on the host (like `add_file`, `remove_file`, etc.). Failure to do so will
-/// cause a hang/deadlock due to Salsa's single-writer, multi-reader model.
-#[derive(Clone)]
-pub struct Analysis {
-    db: IdeDatabase,
-    registry: Arc<RwLock<FileRegistry>>,
-    /// Cached `ProjectFiles` for HIR queries
-    /// This is fetched from the registry when the snapshot is created
-    project_files: Option<graphql_base_db::ProjectFiles>,
-}
-
-impl Analysis {
-    /// Get diagnostics for a file
-    ///
-    /// Returns syntax errors, validation errors, and lint warnings.
-    pub fn diagnostics(&self, file: &FilePath) -> Vec<Diagnostic> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        let analysis_diagnostics =
-            graphql_analysis::file_diagnostics(&self.db, content, metadata, self.project_files);
-
-        analysis_diagnostics
-            .iter()
-            .map(convert_diagnostic)
-            .collect()
-    }
-
-    /// Get diagnostics for all files affected by a change to `changed_file`.
-    ///
-    /// Always includes diagnostics for the changed file itself. Additionally:
-    /// - If the changed file is a **schema** file, re-validates all document files
-    ///   (every operation/fragment validates against the schema).
-    /// - If the changed file is a **document** file containing fragments,
-    ///   re-validates files that spread those fragments.
-    /// - If the changed file has named operations, re-validates files with
-    ///   same-named operations (uniqueness checks).
-    ///
-    /// Salsa memoization ensures unaffected files return cached results instantly.
-    pub fn diagnostics_for_change(
-        &self,
-        changed_file: &FilePath,
-    ) -> HashMap<FilePath, Vec<Diagnostic>> {
-        let mut result = HashMap::new();
-
-        // Always include diagnostics for the changed file
-        result.insert(changed_file.clone(), self.diagnostics(changed_file));
-
-        let Some(project_files) = self.project_files else {
-            return result;
-        };
-
-        let (is_schema, is_document, changed_file_id) = {
-            let registry = self.registry.read();
-            let Some(file_id) = registry.get_file_id(changed_file) else {
-                return result;
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return result;
-            };
-            (
-                metadata.is_schema(&self.db),
-                metadata.is_document(&self.db),
-                file_id,
-            )
-        };
-
-        if is_schema {
-            // Schema change: re-validate all document files
-            let document_files: Vec<FilePath> = {
-                let registry = self.registry.read();
-                registry
-                    .all_file_ids()
-                    .into_iter()
-                    .filter(|&id| {
-                        registry
-                            .get_metadata(id)
-                            .is_some_and(|m| m.is_document(&self.db))
-                    })
-                    .filter_map(|id| registry.get_path(id))
-                    .collect()
-            };
-
-            for doc_file in document_files {
-                if doc_file != *changed_file {
-                    result.insert(doc_file.clone(), self.diagnostics(&doc_file));
-                }
-            }
-        } else if is_document {
-            // Document file change: find affected files via fragment/operation dependencies
-            let affected = self.find_affected_document_files(changed_file_id, project_files);
-            for affected_file in affected {
-                if affected_file != *changed_file {
-                    result.insert(affected_file.clone(), self.diagnostics(&affected_file));
-                }
-            }
-        }
-
-        result
-    }
-
-    /// Get all diagnostics (per-file + project-wide lints) for files affected by a change.
-    ///
-    /// Combines [`diagnostics_for_change`] with [`project_lint_diagnostics`] into a single
-    /// merged result. Each file in the result has its full set of diagnostics:
-    /// per-file validation, per-file lints, and project-wide lints.
-    ///
-    /// Use this from `did_save` to publish complete diagnostics in one pass.
-    pub fn all_diagnostics_for_change(
-        &self,
-        changed_file: &FilePath,
-    ) -> HashMap<FilePath, Vec<Diagnostic>> {
-        let mut result = self.diagnostics_for_change(changed_file);
-
-        // Merge project-wide lint diagnostics into the result
-        let project_diagnostics = self.project_lint_diagnostics();
-        for (file_path, diagnostics) in project_diagnostics {
-            result.entry(file_path).or_default().extend(diagnostics);
-        }
-
-        result
-    }
-
-    /// Find document files affected by a change to the given document file.
-    ///
-    /// Returns files that:
-    /// 1. Spread fragments defined in the changed file (directly or transitively)
-    /// 2. Have same-named operations as the changed file (uniqueness checks)
-    fn find_affected_document_files(
-        &self,
-        changed_file_id: graphql_base_db::FileId,
-        project_files: graphql_base_db::ProjectFiles,
-    ) -> Vec<FilePath> {
-        let registry = self.registry.read();
-
-        // Get fragments and operations defined in the changed file
-        let (content, metadata) = {
-            let Some(content) = registry.get_content(changed_file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(changed_file_id) else {
-                return Vec::new();
-            };
-            (content, metadata)
-        };
-
-        let changed_fragments =
-            graphql_hir::file_fragments(&self.db, changed_file_id, content, metadata);
-        let changed_operations =
-            graphql_hir::file_operations(&self.db, changed_file_id, content, metadata);
-
-        // If no fragments or named operations, no cross-file dependencies
-        let has_fragments = !changed_fragments.is_empty();
-        let has_named_ops = changed_operations.iter().any(|op| op.name.is_some());
-        if !has_fragments && !has_named_ops {
-            return Vec::new();
-        }
-
-        // Collect fragment names from this file
-        let fragment_names: std::collections::HashSet<std::sync::Arc<str>> =
-            changed_fragments.iter().map(|f| f.name.clone()).collect();
-
-        // Collect named operation names from this file
-        let operation_names: std::collections::HashSet<std::sync::Arc<str>> = changed_operations
-            .iter()
-            .filter_map(|op| op.name.clone())
-            .collect();
-
-        // Build the set of fragment names we care about:
-        // 1. Fragments currently defined in this file
-        // 2. Fragments that transitively spread our fragments
-        let mut affected_fragment_names = fragment_names.clone();
-        let all_fragments_index = if has_fragments {
-            let index = graphql_hir::project_fragment_name_index(&self.db, project_files);
-            let spreads_index = graphql_hir::fragment_spreads_index(&self.db, project_files);
-            // Walk reverse: find fragments that spread OUR fragments
-            for (frag_name, spreads) in spreads_index.iter() {
-                if spreads.iter().any(|s| fragment_names.contains(s)) {
-                    affected_fragment_names.insert(frag_name.clone());
-                }
-            }
-            Some(index)
-        } else {
-            None
-        };
-
-        let mut affected_files = Vec::new();
-        let doc_ids = project_files.document_file_ids(&self.db).ids(&self.db);
-
-        for file_id in doc_ids.iter() {
-            if *file_id == changed_file_id {
-                continue;
-            }
-
-            let Some((file_content, file_metadata)) =
-                graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-            else {
-                continue;
-            };
-
-            let mut is_affected = false;
-
-            if has_fragments {
-                let used_names = graphql_hir::file_used_fragment_names(
-                    &self.db,
-                    *file_id,
-                    file_content,
-                    file_metadata,
-                );
-
-                // Check 1: does this file spread any of our current/transitive fragments?
-                if used_names
-                    .iter()
-                    .any(|n| affected_fragment_names.contains(n))
-                {
-                    is_affected = true;
-                }
-
-                // Check 2: does this file spread a fragment that no longer exists?
-                // This catches the rename/delete case: the old fragment name is gone
-                // from the project index, so any file still referencing it needs refresh.
-                if !is_affected {
-                    if let Some(ref index) = all_fragments_index {
-                        if used_names.iter().any(|n| !index.contains_key(n)) {
-                            is_affected = true;
-                        }
-                    }
-                }
-            }
-
-            // Check if this file has same-named operations (uniqueness)
-            if !is_affected && has_named_ops {
-                let file_op_names = graphql_hir::file_operation_names(
-                    &self.db,
-                    *file_id,
-                    file_content,
-                    file_metadata,
-                );
-                if file_op_names
-                    .iter()
-                    .any(|info| operation_names.contains(&info.name))
-                {
-                    is_affected = true;
-                }
-            }
-
-            if is_affected {
-                if let Some(path) = registry.get_path(*file_id) {
-                    affected_files.push(path);
-                }
-            }
-        }
-
-        affected_files
-    }
-
-    /// Get only validation diagnostics for a file (excludes custom lint rules)
-    ///
-    /// Returns only GraphQL spec validation errors, not custom lint rule violations.
-    /// Use this for the `validate` command to avoid duplicating lint checks.
-    pub fn validation_diagnostics(&self, file: &FilePath) -> Vec<Diagnostic> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        let analysis_diagnostics = graphql_analysis::file_validation_diagnostics(
-            &self.db,
-            content,
-            metadata,
-            self.project_files,
-        );
-
-        analysis_diagnostics
-            .iter()
-            .map(convert_diagnostic)
-            .collect()
-    }
-
-    /// Get only lint diagnostics for a file (excludes validation errors)
-    ///
-    /// Returns only custom lint rule violations, not GraphQL spec validation errors.
-    pub fn lint_diagnostics(&self, file: &FilePath) -> Vec<Diagnostic> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        let lint_diagnostics = graphql_analysis::lint_integration::lint_file(
-            &self.db,
-            content,
-            metadata,
-            self.project_files,
-        );
-
-        lint_diagnostics.iter().map(convert_diagnostic).collect()
-    }
-
-    /// Get semantic tokens for a file
-    ///
-    /// Returns tokens for syntax highlighting with semantic information,
-    /// including deprecation status for fields.
-    pub fn semantic_tokens(&self, file: &FilePath) -> Vec<SemanticToken> {
-        let registry = self.registry.read();
-        semantic_tokens::semantic_tokens(&self.db, &registry, self.project_files, file)
-    }
-
-    /// Get folding ranges for a file
-    ///
-    /// Returns foldable regions for:
-    /// - Operation definitions (query, mutation, subscription)
-    /// - Fragment definitions
-    /// - Selection sets
-    /// - Block comments
-    pub fn folding_ranges(&self, file: &FilePath) -> Vec<FoldingRange> {
-        let registry = self.registry.read();
-        folding_ranges::folding_ranges(&self.db, &registry, file)
-    }
-
-    /// Get inlay hints for a file within an optional range.
-    ///
-    /// Returns inlay hints showing return types after scalar field selections.
-    /// Includes support for the `__typename` introspection field.
-    ///
-    /// If `range` is provided, only returns hints within that range for efficiency.
-    pub fn inlay_hints(&self, file: &FilePath, range: Option<Range>) -> Vec<InlayHint> {
-        let registry = self.registry.read();
-        inlay_hints::inlay_hints(&self.db, &registry, self.project_files, file, range)
-    }
-
-    /// Get project-wide lint diagnostics (e.g., unused fields, unique names)
-    ///
-    /// Returns a map of file paths -> diagnostics for project-wide lint rules.
-    /// These are expensive rules that analyze the entire project.
-    pub fn project_lint_diagnostics(&self) -> HashMap<FilePath, Vec<Diagnostic>> {
-        let diagnostics_by_file_id = graphql_analysis::lint_integration::project_lint_diagnostics(
-            &self.db,
-            self.project_files,
-        );
-
-        let mut results = HashMap::new();
-        let registry = self.registry.read();
-
-        for (file_id, diagnostics) in diagnostics_by_file_id.iter() {
-            if let Some(file_path) = registry.get_path(*file_id) {
-                let converted: Vec<Diagnostic> =
-                    diagnostics.iter().map(convert_diagnostic).collect();
-
-                if !converted.is_empty() {
-                    results.insert(file_path, converted);
-                }
-            }
-        }
-
-        results
-    }
-
-    /// Get all diagnostics for all files, merging per-file and project-wide diagnostics
-    ///
-    /// This is a convenience method for publishing diagnostics. It:
-    /// 1. Gets per-file diagnostics (parse errors, validation errors, per-file lint rules)
-    /// 2. Gets project-wide lint diagnostics (unused fields, etc.)
-    /// 3. Merges them per file
-    ///
-    /// Returns a map of file paths -> all diagnostics for that file.
-    pub fn all_diagnostics(&self) -> HashMap<FilePath, Vec<Diagnostic>> {
-        let mut results: HashMap<FilePath, Vec<Diagnostic>> = HashMap::new();
-
-        // Get all registered files
-        let all_file_paths: Vec<FilePath> = {
-            let registry = self.registry.read();
-            registry
-                .all_file_ids()
-                .into_iter()
-                .filter_map(|file_id| registry.get_path(file_id))
-                .collect()
-        };
-
-        // Get per-file diagnostics for all files
-        for file_path in &all_file_paths {
-            let per_file = self.diagnostics(file_path);
-            if !per_file.is_empty() {
-                results.insert(file_path.clone(), per_file);
-            }
-        }
-
-        // Get project-wide diagnostics and merge
-        let project_diagnostics = self.project_lint_diagnostics();
-        for (file_path, diagnostics) in project_diagnostics {
-            results.entry(file_path).or_default().extend(diagnostics);
-        }
-
-        results
-    }
-
-    /// Get all diagnostics for a single file, merging per-file and project-wide diagnostics
-    ///
-    /// This returns the complete set of diagnostics for a file:
-    /// - Per-file diagnostics (parse errors, validation errors, per-file lint rules)
-    /// - Project-wide diagnostics (`unused_fields`, etc.) that apply to this file
-    ///
-    /// Use this when publishing diagnostics to avoid overwriting project-wide diagnostics
-    /// with only per-file diagnostics.
-    pub fn all_diagnostics_for_file(&self, file: &FilePath) -> Vec<Diagnostic> {
-        let mut results = self.diagnostics(file);
-
-        // Add project-wide diagnostics for this file
-        let project_diagnostics = self.project_lint_diagnostics();
-        if let Some(project_diags) = project_diagnostics.get(file) {
-            results.extend(project_diags.iter().cloned());
-        }
-
-        results
-    }
-
-    /// Get all diagnostics for a specific set of files, merging per-file and project-wide diagnostics
-    ///
-    /// This is useful when you want diagnostics for specific files (e.g., loaded document files)
-    /// rather than all files in the registry.
-    pub fn all_diagnostics_for_files(
-        &self,
-        files: &[FilePath],
-    ) -> HashMap<FilePath, Vec<Diagnostic>> {
-        let mut results: HashMap<FilePath, Vec<Diagnostic>> = HashMap::new();
-
-        // Get per-file diagnostics for specified files
-        for file_path in files {
-            let per_file = self.diagnostics(file_path);
-            if !per_file.is_empty() {
-                results.insert(file_path.clone(), per_file);
-            }
-        }
-
-        // Get project-wide diagnostics and merge
-        let project_diagnostics = self.project_lint_diagnostics();
-        for (file_path, diagnostics) in project_diagnostics {
-            // Only include if the file is in our set OR it's a schema file with issues
-            // (project-wide lints like unused_fields report on schema files)
-            results.entry(file_path).or_default().extend(diagnostics);
-        }
-
-        results
-    }
-
-    /// Get raw lint diagnostics with fix information for a file
-    ///
-    /// Returns `LintDiagnostic` objects that include fix information.
-    /// Use this for implementing auto-fix functionality.
-    pub fn lint_diagnostics_with_fixes(
-        &self,
-        file: &FilePath,
-    ) -> Vec<graphql_linter::LintDiagnostic> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        graphql_analysis::lint_integration::lint_file_with_fixes(
-            &self.db,
-            content,
-            metadata,
-            self.project_files,
-        )
-    }
-
-    /// Get project-wide raw lint diagnostics with fix information
-    ///
-    /// Returns a map of file paths -> `LintDiagnostic` objects that include fix information.
-    pub fn project_lint_diagnostics_with_fixes(
-        &self,
-    ) -> HashMap<FilePath, Vec<graphql_linter::LintDiagnostic>> {
-        let diagnostics_by_file_id =
-            graphql_analysis::lint_integration::project_lint_diagnostics_with_fixes(
-                &self.db,
-                self.project_files,
-            );
-
-        let mut results = HashMap::new();
-        let registry = self.registry.read();
-
-        for (file_id, diagnostics) in diagnostics_by_file_id {
-            if let Some(file_path) = registry.get_path(file_id) {
-                if !diagnostics.is_empty() {
-                    results.insert(file_path, diagnostics);
-                }
-            }
-        }
-
-        results
-    }
-
-    /// Get the content of a file
-    ///
-    /// Returns the text content of the file if it exists in the registry.
-    pub fn file_content(&self, file: &FilePath) -> Option<String> {
-        let registry = self.registry.read();
-        let file_id = registry.get_file_id(file)?;
-        let content = registry.get_content(file_id)?;
-        Some(content.text(&self.db).to_string())
-    }
-
-    /// Get the status of the project (file counts, schema loaded, etc.)
-    ///
-    /// Returns status information for the LSP status command.
-    #[must_use]
-    pub fn project_status(&self) -> ProjectStatus {
-        let Some(project_files) = self.project_files else {
-            return ProjectStatus::default();
-        };
-
-        let schema_file_count = project_files.schema_file_ids(&self.db).ids(&self.db).len();
-        let document_file_count = project_files
-            .document_file_ids(&self.db)
-            .ids(&self.db)
-            .len();
-
-        ProjectStatus::new(schema_file_count, document_file_count)
-    }
-
-    /// Get field usage coverage report for the project
-    ///
-    /// Analyzes which schema fields are used in operations and returns
-    /// detailed coverage statistics. This is useful for understanding
-    /// schema usage patterns and finding unused fields.
-    pub fn field_coverage(&self) -> Option<FieldCoverageReport> {
-        let pf = self.project_files?;
-        Some(FieldCoverageReport::from(
-            graphql_analysis::analyze_field_usage(&self.db, pf),
-        ))
-    }
-
-    /// Get field usage for a specific field
-    ///
-    /// Returns usage information for a field if it exists in the schema.
-    /// Useful for enhancing hover to show "Used in N operations".
-    pub fn field_usage(&self, type_name: &str, field_name: &str) -> Option<FieldUsageInfo> {
-        let pf = self.project_files?;
-        let coverage = graphql_analysis::analyze_field_usage(&self.db, pf);
-        let key = (
-            std::sync::Arc::from(type_name),
-            std::sync::Arc::from(field_name),
-        );
-        coverage.field_usages.get(&key).map(|usage| FieldUsageInfo {
-            usage_count: usage.usage_count,
-            operations: usage.operations.iter().map(ToString::to_string).collect(),
-        })
-    }
-
-    /// Get complexity analysis for all operations in the project
-    ///
-    /// Analyzes each operation's selection set to calculate:
-    /// - Total complexity score (with list multipliers)
-    /// - Maximum depth
-    /// - Per-field complexity breakdown
-    /// - Connection pattern detection (Relay-style edges/nodes/pageInfo)
-    /// - Warnings about potential issues (nested pagination, etc.)
-    pub fn complexity_analysis(&self) -> Vec<ComplexityAnalysis> {
-        let Some(project_files) = self.project_files else {
-            return Vec::new();
-        };
-
-        // Get all operations in the project
-        let operations = graphql_hir::all_operations(&self.db, project_files);
-        let schema_types = graphql_hir::schema_types(&self.db, project_files);
-
-        let mut results = Vec::new();
-
-        for operation in operations.iter() {
-            // Get file information for this operation
-            let registry = self.registry.read();
-            let Some(file_path) = registry.get_path(operation.file_id) else {
-                continue;
-            };
-            let Some(content) = registry.get_content(operation.file_id) else {
-                continue;
-            };
-            let Some(metadata) = registry.get_metadata(operation.file_id) else {
-                continue;
-            };
-            drop(registry);
-
-            // Get operation body
-            let body = graphql_hir::operation_body(&self.db, content, metadata, operation.index);
-
-            // Get operation location for the range
-            let range = if let Some(ref name) = operation.name {
-                let parse = graphql_syntax::parse(&self.db, content, metadata);
-                let mut found_range = None;
-                for doc in parse.documents() {
-                    if let Some(ranges) = find_operation_definition_ranges(doc.tree, name) {
-                        let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                        let doc_line_offset = doc.line_offset;
-                        found_range = Some(adjust_range_for_line_offset(
-                            offset_range_to_range(
-                                &doc_line_index,
-                                ranges.def_start,
-                                ranges.def_end,
-                            ),
-                            doc_line_offset,
-                        ));
-                        break;
-                    }
-                }
-                found_range.unwrap_or_else(|| Range::new(Position::new(0, 0), Position::new(0, 0)))
-            } else {
-                Range::new(Position::new(0, 0), Position::new(0, 0))
-            };
-
-            // Create complexity analysis
-            let op_name = operation
-                .name
-                .as_ref()
-                .map_or_else(|| "<anonymous>".to_string(), ToString::to_string);
-
-            #[allow(clippy::match_same_arms)]
-            let op_type = match operation.operation_type {
-                graphql_hir::OperationType::Query => "query",
-                graphql_hir::OperationType::Mutation => "mutation",
-                graphql_hir::OperationType::Subscription => "subscription",
-                _ => "query", // fallback for future operation types
-            };
-
-            let mut analysis = ComplexityAnalysis::new(op_name, op_type, file_path, range);
-
-            // Get the root type for this operation
-            #[allow(clippy::match_same_arms)]
-            let root_type_name = match operation.operation_type {
-                graphql_hir::OperationType::Query => "Query",
-                graphql_hir::OperationType::Mutation => "Mutation",
-                graphql_hir::OperationType::Subscription => "Subscription",
-                _ => "Query", // fallback for future operation types
-            };
-
-            // Analyze the operation body
-            analyze_selections(
-                &body.selections,
-                schema_types,
-                root_type_name,
-                "",
-                0,
-                1,
-                &mut analysis,
-                false,
-            );
-
-            results.push(analysis);
-        }
-
-        results
-    }
-
-    /// Get completions at a position
-    ///
-    /// Returns a list of completion items appropriate for the context.
-    pub fn completions(&self, file: &FilePath, position: Position) -> Option<Vec<CompletionItem>> {
-        let registry = self.registry.read();
-        completion::completions(&self.db, &registry, self.project_files, file, position)
-    }
-
-    /// Get hover information at a position
-    ///
-    /// Returns documentation, type information, etc.
-    pub fn hover(&self, file: &FilePath, position: Position) -> Option<HoverResult> {
-        let registry = self.registry.read();
-        hover::hover(&self.db, &registry, self.project_files, file, position)
-    }
-
-    /// Get signature help at a position
-    ///
-    /// Returns argument information when inside a field or directive argument list.
-    pub fn signature_help(&self, file: &FilePath, position: Position) -> Option<SignatureHelp> {
-        let registry = self.registry.read();
-        signature_help::signature_help(&self.db, &registry, self.project_files, file, position)
-    }
-
-    /// Get goto definition locations for the symbol at a position
-    ///
-    /// Returns the definition location(s) for types, fields, fragments, etc.
-    pub fn goto_definition(&self, file: &FilePath, position: Position) -> Option<Vec<Location>> {
-        let registry = self.registry.read();
-        goto_definition::goto_definition(&self.db, &registry, self.project_files, file, position)
-    }
-
-    /// Find all references to the symbol at a position
-    ///
-    /// Returns locations of all usages of types, fields, fragments, etc.
-    pub fn find_references(
-        &self,
-        file: &FilePath,
-        position: Position,
-        include_declaration: bool,
-    ) -> Option<Vec<Location>> {
-        let registry = self.registry.read();
-        references::find_references(
-            &self.db,
-            &registry,
-            self.project_files,
-            file,
-            position,
-            include_declaration,
-        )
-    }
-
-    /// Check if the symbol at a position can be renamed, returning its range.
-    pub fn prepare_rename(&self, file: &FilePath, position: Position) -> Option<Range> {
-        let registry = self.registry.read();
-        rename::prepare_rename(&self.db, &registry, file, position)
-    }
-
-    /// Rename the symbol at a position to a new name.
-    pub fn rename(
-        &self,
-        file: &FilePath,
-        position: Position,
-        new_name: &str,
-    ) -> Option<RenameResult> {
-        let registry = self.registry.read();
-        rename::rename(
-            &self.db,
-            &registry,
-            self.project_files,
-            file,
-            position,
-            new_name,
-        )
-    }
-
-    /// Find all references to a fragment
-    pub fn find_fragment_references(
-        &self,
-        fragment_name: &str,
-        include_declaration: bool,
-    ) -> Vec<Location> {
-        let registry = self.registry.read();
-        references::find_fragment_references(
-            &self.db,
-            &registry,
-            self.project_files,
-            fragment_name,
-            include_declaration,
-        )
-    }
-
-    /// Get selection ranges for smart expand/shrink selection
-    ///
-    /// Returns a `SelectionRange` for each input position, forming a linked list
-    /// from the innermost syntax element to the outermost (document).
-    /// This powers the "Expand Selection" (Shift+Alt+Right) and
-    /// "Shrink Selection" (Shift+Alt+Left) features.
-    pub fn selection_ranges(
-        &self,
-        file: &FilePath,
-        positions: &[Position],
-    ) -> Vec<Option<SelectionRange>> {
-        let registry = self.registry.read();
-        selection_range::selection_ranges(&self.db, &registry, file, positions)
-    }
-
-    /// Get code lenses for deprecated fields in a schema file
-    ///
-    /// Returns code lens information for each deprecated field definition,
-    /// including the usage count and locations for navigation.
-    pub fn deprecated_field_code_lenses(&self, file: &FilePath) -> Vec<CodeLensInfo> {
-        let registry = self.registry.read();
-        code_lenses::deprecated_field_code_lenses(&self.db, &registry, self.project_files, file)
-    }
-
-    /// Get document symbols for a file (hierarchical outline)
-    ///
-    /// Returns types, operations, and fragments with their fields as children.
-    /// This powers the "Go to Symbol in Editor" (Cmd+Shift+O) feature.
-    pub fn document_symbols(&self, file: &FilePath) -> Vec<DocumentSymbol> {
-        let registry = self.registry.read();
-        symbols::document_symbols(&self.db, &registry, file)
-    }
-
-    /// Search for workspace symbols matching a query
-    ///
-    /// Returns matching types, operations, and fragments across all files.
-    /// This powers the "Go to Symbol in Workspace" (Cmd+T) feature.
-    pub fn workspace_symbols(&self, query: &str) -> Vec<WorkspaceSymbol> {
-        let registry = self.registry.read();
-        symbols::workspace_symbols(&self.db, &registry, self.project_files, query)
-    }
-
-    /// Get schema statistics
-    ///
-    /// Returns counts of types by kind, total fields, and directives.
-    /// This uses the HIR layer directly for accurate field counting.
-    pub fn schema_stats(&self) -> SchemaStats {
-        let Some(project_files) = self.project_files else {
-            return SchemaStats::default();
-        };
-
-        let types = graphql_hir::schema_types(&self.db, project_files);
-        let mut stats = SchemaStats::default();
-
-        for type_def in types.values() {
-            match type_def.kind {
-                graphql_hir::TypeDefKind::Object => stats.objects += 1,
-                graphql_hir::TypeDefKind::Interface => stats.interfaces += 1,
-                graphql_hir::TypeDefKind::Union => stats.unions += 1,
-                graphql_hir::TypeDefKind::Enum => stats.enums += 1,
-                graphql_hir::TypeDefKind::Scalar => stats.scalars += 1,
-                graphql_hir::TypeDefKind::InputObject => stats.input_objects += 1,
-                _ => {} // ignore future type kinds
-            }
-            // Count fields for types that have fields
-            stats.total_fields += type_def.fields.len();
-        }
-
-        // Count directive definitions from schema files (excluding built-ins)
-        let schema_ids = project_files.schema_file_ids(&self.db).ids(&self.db);
-        for file_id in schema_ids.iter() {
-            let Some((content, metadata)) =
-                graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-            else {
-                continue;
-            };
-
-            // Skip built-in directive files
-            let registry = self.registry.read();
-            if let Some(path) = registry.get_path(*file_id) {
-                let path_str = path.as_str();
-                if path_str == "client_builtins.graphql" || path_str == "schema_builtins.graphql" {
-                    drop(registry);
-                    continue;
-                }
-            }
-            drop(registry);
-
-            let parse = graphql_syntax::parse(&self.db, content, metadata);
-            // Count directive definitions by checking if the definition is a directive
-            // Directives in GraphQL SDL start with "directive @"
-            for doc in parse.documents() {
-                for definition in &doc.ast.definitions {
-                    if definition.as_directive_definition().is_some() {
-                        stats.directives += 1;
-                    }
-                }
-            }
-        }
-
-        stats
-    }
-
-    /// Get fragment usage analysis for the project
-    ///
-    /// Returns information about each fragment: its definition location,
-    /// all usages (fragment spreads), and transitive dependencies.
-    pub fn fragment_usages(&self) -> Vec<FragmentUsage> {
-        let Some(project_files) = self.project_files else {
-            return Vec::new();
-        };
-
-        let fragments = graphql_hir::all_fragments(&self.db, project_files);
-        let mut results = Vec::new();
-
-        for (name, fragment) in fragments {
-            // Get definition location
-            let Some((def_file, def_range)) = self.get_fragment_def_info(fragment) else {
-                continue;
-            };
-
-            // Get all usages (fragment spreads) excluding the definition
-            let spread_locations = self.find_fragment_references(name, false);
-            let usages: Vec<FragmentReference> = spread_locations
-                .into_iter()
-                .map(FragmentReference::new)
-                .collect();
-
-            // Get transitive dependencies using the fragment spreads index
-            let transitive_deps = self.compute_transitive_dependencies(name, project_files);
-
-            results.push(FragmentUsage {
-                name: name.to_string(),
-                definition_file: def_file,
-                definition_range: def_range,
-                usages,
-                transitive_dependencies: transitive_deps,
-            });
-        }
-
-        // Sort by name for consistent ordering
-        results.sort_by(|a, b| a.name.cmp(&b.name));
-        results
-    }
-
-    /// Get fragment definition file and range
-    fn get_fragment_def_info(
-        &self,
-        fragment: &graphql_hir::FragmentStructure,
-    ) -> Option<(FilePath, Range)> {
-        let registry = self.registry.read();
-        let file_path = registry.get_path(fragment.file_id)?;
-        let content = registry.get_content(fragment.file_id)?;
-        let metadata = registry.get_metadata(fragment.file_id)?;
-        drop(registry);
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        for doc in parse.documents() {
-            if let Some(ranges) = find_fragment_definition_full_range(doc.tree, &fragment.name) {
-                let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                let range = adjust_range_for_line_offset(
-                    offset_range_to_range(&doc_line_index, ranges.name_start, ranges.name_end),
-                    doc.line_offset,
-                );
-                return Some((file_path, range));
-            }
-        }
-
-        None
-    }
-
-    /// Compute transitive fragment dependencies
-    fn compute_transitive_dependencies(
-        &self,
-        fragment_name: &str,
-        project_files: graphql_base_db::ProjectFiles,
-    ) -> Vec<String> {
-        let spreads_index = graphql_hir::fragment_spreads_index(&self.db, project_files);
-
-        let mut visited = std::collections::HashSet::new();
-        let mut to_visit = Vec::new();
-
-        // Start with direct dependencies
-        if let Some(direct_deps) = spreads_index.get(fragment_name) {
-            to_visit.extend(direct_deps.iter().cloned());
-        }
-
-        while let Some(dep_name) = to_visit.pop() {
-            if !visited.insert(dep_name.clone()) {
-                continue; // Already visited (handles cycles)
-            }
-
-            // Add transitive dependencies
-            if let Some(nested_deps) = spreads_index.get(&dep_name) {
-                for nested in nested_deps {
-                    if !visited.contains(nested) {
-                        to_visit.push(nested.clone());
-                    }
-                }
-            }
-        }
-
-        let mut deps: Vec<String> = visited.into_iter().map(|s| s.to_string()).collect();
-        deps.sort();
-        deps
-    }
-
-    /// Get code lenses for a file
-    ///
-    /// Returns code lenses for fragment definitions showing reference counts.
-    pub fn code_lenses(&self, file: &FilePath) -> Vec<CodeLens> {
-        let fragment_usages = self.fragment_usages();
-        let registry = self.registry.read();
-        code_lenses::code_lenses(
-            &self.db,
-            &registry,
-            self.project_files,
-            file,
-            &fragment_usages,
-        )
-    }
-}
-
-// Helper functions are now in feature modules
-
-/// Analyze selections recursively to calculate complexity
-#[allow(clippy::too_many_arguments)]
-fn analyze_selections(
-    selections: &[graphql_hir::Selection],
-    schema_types: &std::collections::HashMap<Arc<str>, graphql_hir::TypeDef>,
-    parent_type_name: &str,
-    path_prefix: &str,
-    depth: u32,
-    multiplier: u32,
-    analysis: &mut ComplexityAnalysis,
-    in_connection: bool,
-) {
-    // Update max depth
-    if depth > analysis.depth {
-        analysis.depth = depth;
-    }
-
-    for selection in selections {
-        match selection {
-            graphql_hir::Selection::Field {
-                name,
-                selection_set,
-                ..
-            } => {
-                let field_name = name.to_string();
-                let path = if path_prefix.is_empty() {
-                    field_name.clone()
-                } else {
-                    format!("{path_prefix}.{field_name}")
-                };
-
-                // Get field type info from schema
-                let (is_list, inner_type_name) =
-                    get_type_info(schema_types, parent_type_name, &field_name);
-
-                // Calculate field multiplier
-                let field_multiplier = if is_list {
-                    multiplier * 10 // Default list multiplier
-                } else {
-                    multiplier
-                };
-
-                // Check for connection pattern
-                let field_is_connection =
-                    is_connection_pattern(&field_name, schema_types, &inner_type_name);
-
-                // Warn about nested pagination
-                if in_connection && field_is_connection {
-                    analysis.warnings.push(format!(
-                        "Nested pagination detected at {path}. This can cause performance issues."
-                    ));
-                }
-
-                // Calculate complexity for this field
-                let field_complexity = field_multiplier;
-                analysis.total_complexity += field_complexity;
-
-                // Add to breakdown
-                let mut fc = FieldComplexity::new(&path, &field_name, field_complexity)
-                    .with_multiplier(if is_list { 10 } else { 1 })
-                    .with_depth(depth)
-                    .with_connection(field_is_connection);
-
-                if in_connection && field_is_connection {
-                    fc = fc.with_warning("Nested pagination");
-                }
-
-                analysis.breakdown.push(fc);
-
-                // Recurse into nested selections
-                if !selection_set.is_empty() {
-                    analyze_selections(
-                        selection_set,
-                        schema_types,
-                        &inner_type_name,
-                        &path,
-                        depth + 1,
-                        field_multiplier,
-                        analysis,
-                        field_is_connection || in_connection,
-                    );
-                }
-            }
-            graphql_hir::Selection::FragmentSpread { .. }
-            | graphql_hir::Selection::InlineFragment { .. } => {
-                // For simplicity, we don't deeply analyze fragment spreads in this implementation
-                // A full implementation would resolve the fragment and analyze its selections
-            }
-        }
-    }
-}
-
-/// Check if a field follows the Relay connection pattern (edges/nodes/pageInfo)
-fn is_connection_pattern(
-    _field_name: &str,
-    schema_types: &std::collections::HashMap<Arc<str>, graphql_hir::TypeDef>,
-    type_name: &str,
-) -> bool {
-    // Check if the return type has edges, nodes, or pageInfo fields
-    if let Some(type_def) = schema_types.get(type_name) {
-        if type_def.kind == graphql_hir::TypeDefKind::Object {
-            let has_edges = type_def.fields.iter().any(|f| f.name.as_ref() == "edges");
-            let has_page_info = type_def
-                .fields
-                .iter()
-                .any(|f| f.name.as_ref() == "pageInfo");
-            let has_nodes = type_def.fields.iter().any(|f| f.name.as_ref() == "nodes");
-
-            return (has_edges || has_nodes) && has_page_info;
-        }
-    }
-    false
-}
-
-/// Get type information for a field: (`is_list`, `inner_type_name`)
-fn get_type_info(
-    schema_types: &std::collections::HashMap<Arc<str>, graphql_hir::TypeDef>,
-    parent_type_name: &str,
-    field_name: &str,
-) -> (bool, String) {
-    if let Some(type_def) = schema_types.get(parent_type_name) {
-        if type_def.kind == graphql_hir::TypeDefKind::Object {
-            if let Some(field) = type_def
-                .fields
-                .iter()
-                .find(|f| f.name.as_ref() == field_name)
-            {
-                let type_ref = &field.type_ref;
-                return (type_ref.is_list, type_ref.name.to_string());
-            }
-        }
-    }
-    (false, "Unknown".to_string())
-}
-
 #[cfg(test)]
 #[allow(clippy::needless_raw_string_hashes)]
 mod tests {
+    use std::collections::HashMap;
+
     use super::*;
-    use crate::helpers::{convert_position, convert_range, convert_severity, position_to_offset};
+    use crate::helpers::{
+        convert_diagnostic, convert_position, convert_range, convert_severity, position_to_offset,
+    };
 
     #[test]
     fn test_analysis_host_creation() {
